extract save method
