+++ saleor/graphql/account/mutations/base.py
@@ -1,19 +1,15 @@
 from collections import defaultdict
-from urllib.parse import urlencode
 
 import graphene
-from django.contrib.auth.tokens import default_token_generator
 from django.core.exceptions import ValidationError
 
 from ....account import events as account_events
 from ....account.error_codes import AccountErrorCode
-from ....account.notifications import send_set_password_notification
 from ....account.search import prepare_user_search_document_value
 from ....checkout import AddressType
 from ....core.exceptions import PermissionDenied
-from ....core.tracing import traced_atomic_transaction
 from ....core.utils import metadata_manager
-from ....core.utils.url import prepare_url, validate_storefront_url
+from ....core.utils.url import validate_storefront_url
 from ....giftcard.search import mark_gift_cards_search_index_as_dirty
 from ....giftcard.utils import get_user_gift_cards
 from ....graphql.utils import get_user_or_app_from_context
@@ -22,7 +18,6 @@ from ....permission.enums import AccountPermissions
 from ...account.i18n import I18nMixin
 from ...account.types import Address, AddressInput, User
 from ...app.dataloaders import get_app_promise
-from ...channel.utils import clean_channel, validate_channel
 from ...core import ResolveInfo, SaleorContext
 from ...core.descriptions import (
     DEPRECATED_IN_3X_INPUT,
@@ -347,62 +342,6 @@ class BaseCustomerCreate(DeprecatedModelMutation, I18nMixin):
             cleaned_input["is_confirmed"] = False
 
         return cleaned_input
-
-    @classmethod
-    @traced_atomic_transaction()
-    def save(cls, info: ResolveInfo, instance, cleaned_input):
-        default_shipping_address = cleaned_input.get(SHIPPING_ADDRESS_FIELD)
-        manager = get_plugin_manager_promise(info.context).get()
-        if default_shipping_address:
-            default_shipping_address.save()
-            instance.default_shipping_address = default_shipping_address
-        default_billing_address = cleaned_input.get(BILLING_ADDRESS_FIELD)
-        if default_billing_address:
-            default_billing_address.save()
-            instance.default_billing_address = default_billing_address
-
-        is_creation = instance.pk is None
-        super().save(info, instance, cleaned_input)
-        if default_billing_address:
-            instance.addresses.add(default_billing_address)
-        if default_shipping_address:
-            instance.addresses.add(default_shipping_address)
-
-        instance.search_document = prepare_user_search_document_value(instance)
-        instance.save(update_fields=["search_document", "updated_at"])
-
-        # The instance is a new object in db, create an event
-        if is_creation:
-            cls.call_event(manager.customer_created, instance)
-            account_events.customer_account_created_event(user=instance)
-        else:
-            cls.call_event(manager.customer_updated, instance)
-
-        if redirect_url := cleaned_input.get("redirect_url"):
-            channel_slug = cleaned_input.get("channel")
-            if not instance.is_staff:
-                channel_slug = clean_channel(
-                    channel_slug, error_class=AccountErrorCode, allow_replica=False
-                ).slug
-            elif channel_slug is not None:
-                channel_slug = validate_channel(
-                    channel_slug, error_class=AccountErrorCode
-                ).slug
-            send_set_password_notification(
-                redirect_url,
-                instance,
-                manager,
-                channel_slug,
-            )
-            token = default_token_generator.make_token(instance)
-            params = urlencode({"email": instance.email, "token": token})
-            cls.call_event(
-                manager.account_set_password_requested,
-                instance,
-                channel_slug,
-                token,
-                prepare_url(params, redirect_url),
-            )
 
     @classmethod
     def post_save_action(cls, info: ResolveInfo, instance, cleaned_input):
