+++ saleor/graphql/account/mutations/staff/customer_create.py
@@ -1,11 +1,24 @@
+from urllib.parse import urlencode
+
+from django.contrib.auth.tokens import default_token_generator
+
+from .....account import events as account_events
 from .....account import models
+from .....account.error_codes import AccountErrorCode
+from .....account.notifications import send_set_password_notification
+from .....account.search import prepare_user_search_document_value
+from .....core.tracing import traced_atomic_transaction
+from .....core.utils.url import prepare_url
 from .....permission.enums import AccountPermissions
 from .....webhook.event_types import WebhookEventAsyncType
 from ....account.types import User
+from ....channel.utils import clean_channel, validate_channel
+from ....core import ResolveInfo
 from ....core.doc_category import DOC_CATEGORY_USERS
 from ....core.types import AccountError
 from ....core.utils import WebhookEventInfo
-from ..base import BaseCustomerCreate
+from ....plugins.dataloaders import get_plugin_manager_promise
+from ..base import BILLING_ADDRESS_FIELD, SHIPPING_ADDRESS_FIELD, BaseCustomerCreate
 
 
 class CustomerCreate(BaseCustomerCreate):
@@ -38,3 +51,59 @@ class CustomerCreate(BaseCustomerCreate):
                 description="Setting a new password for the account is requested.",
             ),
         ]
+
+    @classmethod
+    @traced_atomic_transaction()
+    def save(cls, info: ResolveInfo, instance, cleaned_input):
+        default_shipping_address = cleaned_input.get(SHIPPING_ADDRESS_FIELD)
+        manager = get_plugin_manager_promise(info.context).get()
+        if default_shipping_address:
+            default_shipping_address.save()
+            instance.default_shipping_address = default_shipping_address
+        default_billing_address = cleaned_input.get(BILLING_ADDRESS_FIELD)
+        if default_billing_address:
+            default_billing_address.save()
+            instance.default_billing_address = default_billing_address
+
+        is_creation = instance.pk is None
+        super().save(info, instance, cleaned_input)
+        if default_billing_address:
+            instance.addresses.add(default_billing_address)
+        if default_shipping_address:
+            instance.addresses.add(default_shipping_address)
+
+        instance.search_document = prepare_user_search_document_value(instance)
+        instance.save(update_fields=["search_document", "updated_at"])
+
+        # The instance is a new object in db, create an event
+        if is_creation:
+            cls.call_event(manager.customer_created, instance)
+            account_events.customer_account_created_event(user=instance)
+        else:
+            cls.call_event(manager.customer_updated, instance)
+
+        if redirect_url := cleaned_input.get("redirect_url"):
+            channel_slug = cleaned_input.get("channel")
+            if not instance.is_staff:
+                channel_slug = clean_channel(
+                    channel_slug, error_class=AccountErrorCode, allow_replica=False
+                ).slug
+            elif channel_slug is not None:
+                channel_slug = validate_channel(
+                    channel_slug, error_class=AccountErrorCode
+                ).slug
+            send_set_password_notification(
+                redirect_url,
+                instance,
+                manager,
+                channel_slug,
+            )
+            token = default_token_generator.make_token(instance)
+            params = urlencode({"email": instance.email, "token": token})
+            cls.call_event(
+                manager.account_set_password_requested,
+                instance,
+                channel_slug,
+                token,
+                prepare_url(params, redirect_url),
+            )
